small update before branching off to modify Player Input to handle Prompts by implementing it like a FSM to modify the input at runtime as needed.
