+++ Source/UnlimitedEngine/Core/StateIdentifiers.hpp
@@ -15,6 +15,7 @@ namespace States
         Settings,
         GameOver,
         Loading,
+        Dialog,
         HighScoreState,
         GetNameState,
         StartLevelState
